+++ app/[locale]/admin/guidance/groups/[groupId]/edit/page.tsx
@@ -1,14 +1,14 @@
 "use client";
 
 import React, { useState } from "react";
-import { Breadcrumb, Breadcrumbs, Link, Button, Checkbox, Form, ListBoxItem } from "react-aria-components";
+import { Breadcrumb, Breadcrumbs, Link, Button, Checkbox } from "react-aria-components";
 import { useTranslations } from "next-intl";
 import { useParams } from "next/navigation";
 
 // Components
 import PageHeader from "@/components/PageHeader";
-import { ContentContainer, LayoutWithPanel, SidebarPanel } from "@/components/Container";
-import { FormSelect, FormInput, CheckboxGroupComponent } from "@/components/Form";
+import { ContentContainer, LayoutWithPanel } from "@/components/Container";
+import { FormInput, CheckboxGroupComponent } from "@/components/Form";
 
 import { routePath } from "@/utils/routes";
 import styles from "./guidanceGroupEdit.module.scss";