+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 ### Added
-
 - `DashboardListItem` component and some basic tests
 - guidance route definitions to `utils/routes.ts`
 - Added Guidance Dashboard page, SCSS file and tests
@@ -8,6 +7,8 @@
 - Added Guidance Group Edit page, SCSS file and tests
 - Added Guidance Text Create page, SCSS file and tests
 - Added Guidance Text Edit page, SCSS file and tests
+- Duplicate templates to add Organization Templates page
+- Add admin feedback options page
 - Add admin email text page with styling and tests
 - Add static page for Org admin projects dashboard
 - Add static update password page
@@ -51,6 +52,8 @@
 - Added test suites for CheckboxGroup and RadioGroup components that seemed to be missing [#503]
 - Interactive form examples to the styleguide showing required and non-required
   states [#503]
+- New DmpTable component that wraps the react-aria table, columns and rows. [#239]
+- Created the static page for the org admin user dashboard. [#782]
 
 ### Updated